adding cards entity, very detailed entity, need to finish service and Controller

+++ KaosCtrl_Backend/src/main/java/com/kchauntell/KaosCtrl_Backend/Entity/Boards.java
@@ -36,7 +36,7 @@ public class Boards {
             String description,
             LocalDateTime createdAt,
             LocalDateTime updatedAt,
-            int user_id) {
+            long user_id) {
         this.name = name;
         this.description = description;
         this.createdAt = createdAt;
